Move functions arond
